Change class and loader IDs from ints to longs

+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/classfile/ClassfileParser.java
@@ -444,7 +444,7 @@ public final class ClassfileParser {
 
         if (classDefinitionInfo.isHidden()) {
             assert requestedClassType != null;
-            int futureKlassID = env.getNewKlassId();
+            long futureKlassID = env.getNewKlassId();
             classDefinitionInfo.initKlassID(futureKlassID);
             thisKlassName = env.getNames().getOrCreate(Types.hiddenClassName(requestedClassType, futureKlassID));
             thisKlassType = env.getTypes().fromName(thisKlassName);
